requirements.txt
auto-sklearn==0.15.0
gensim
imblearn
lime
matplotlib==3.6.2
mlflow==2.1.1
nltk
numpy==1.23.5
pandas==1.5.2
pycm==3.7
scikit-learn<0.25,>=0.24.0
seaborn==0.12.2
spacy==3.4.4
torch==1.13.1
transformers==4.25.1
